refactor(core): split TransferLockGuard into its own module
rdma_fetch.rs had grown past the 1k-line limit; the guard and its
loopback release-counting test are a self-contained unit.
